Mocks is just an implementation detail of Tasks.
