+++ plugin-gradle/src/test/java/com/diffplug/gradle/spotless/Tasks.java
@@ -15,10 +15,72 @@
  */
 package com.diffplug.gradle.spotless;
 
+import java.io.File;
+
+import org.gradle.api.file.FileCollection;
+import org.gradle.api.file.FileSystemLocation;
+import org.gradle.api.file.FileType;
+import org.gradle.api.provider.Provider;
+import org.gradle.work.ChangeType;
+import org.gradle.work.FileChange;
+import org.gradle.work.InputChanges;
+
+import com.diffplug.common.collect.Iterables;
+import com.diffplug.spotless.FileSignature;
+
 final class Tasks {
 	private Tasks() {}
 
 	static void execute(SpotlessTask task) throws Exception {
-		task.performAction(Mocks.mockIncrementalTaskInputs(task.getProject().getProjectDir(), task.getTarget()));
+		task.performAction(mockIncrementalTaskInputs(task.getProject().getProjectDir(), task.getTarget()));
+	}
+
+	private static InputChanges mockIncrementalTaskInputs(File rootDir, Iterable<File> target) {
+		return new InputChanges() {
+			@Override
+			public boolean isIncremental() {
+				return false;
+			}
+
+			private Iterable<FileChange> getFileChanges() {
+				return Iterables.transform(target, file -> mockInputFileDetails(rootDir, file));
+			}
+
+			@Override
+			public Iterable<FileChange> getFileChanges(FileCollection arg0) {
+				return getFileChanges();
+			}
+
+			@Override
+			public Iterable<FileChange> getFileChanges(Provider<? extends FileSystemLocation> arg0) {
+				return getFileChanges();
+			}
+		};
+	}
+
+	private static FileChange mockInputFileDetails(File rootDir, File file) {
+		return new FileChange() {
+			@Override
+			public ChangeType getChangeType() {
+				return ChangeType.MODIFIED;
+			}
+
+			@Override
+			public File getFile() {
+				return file;
+			}
+
+			@Override
+			public FileType getFileType() {
+				return FileType.FILE;
+			}
+
+			@Override
+			public String getNormalizedPath() {
+				String rootPath = FileSignature.pathNativeToUnix(rootDir.getAbsolutePath());
+				String absPath = FileSignature.pathNativeToUnix(file.getAbsolutePath());
+				return FileSignature.subpath(rootPath, absPath);
+			}
+		};
 	}
 }
